# swagger.yaml
openapi: 3.0.3
info:
  title: Univote - University Voting System API
  description: |
    Complete API documentation for Univote, a secure university voting system with facial recognition, geofencing, and email notifications.
    
    ## Features
    - 🔐 JWT-based authentication for students and admins
    - 👤 Azure Face API facial verification
    - 📍 Geofencing validation (Bowen University campus)
    - 📧 Automated email notifications (welcome, vote confirmation, results)
    - 🗳️ Real-time voting with eligibility checks
    - 📊 Comprehensive analytics and reporting
    - 🔒 Single session enforcement
    - 📝 Audit logging for all critical operations
    
    ## Authentication Flow
    1. **First Login**: Use default password '1234' to get `first_login` token
    2. **Change Password**: Use first_login token to set new password (triggers welcome email)
    3. **Regular Login**: Login with new password to get `student` token
    4. **Access Protected Routes**: Use student/admin token in Authorization header
    
    ## Rate Limits
    - Authentication endpoints: 5 requests per 15 minutes
    - Voting endpoints: 10 requests per minute
    - Face API endpoints: 20 requests per minute
    - General API: 100 requests per 15 minutes
    
    ## Security
    - Passwords hashed with bcrypt (10 rounds)
    - JWT tokens expire after 24 hours
    - Single active session per user
    - Geofencing: 5000m radius from Bowen University (7.8525, 4.2811)
    
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoints. 
        Token types: 'student', 'admin', 'first_login'
  schemas:
...
    VotingSession:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
          example: "Student Union Executive Elections 2024/2025"
        description:
          type: string
          example: "Annual elections for Student Union Government positions"
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, active, completed]
          example: "active"
        eligibility_criteria:
          type: object
          properties:
            colleges:
              type: array
              items:
                type: string
              example: ["COCCS", "COMSS", "COLAW"]
            departments:
              type: array
              items:
                type: string
              example: ["Computer Science", "Biology"]
            levels:
              type: array
              items:
                type: string
              example: ["100", "200", "300", "400"]
        location:
          type: object
          properties:
            type:
              type: string
              example: "Point"
            coordinates:
              type: array
              items:
                type: number
              example: [4.2811, 7.8525]
            name:
              type: string
              example: "Bowen University Campus"
            radius_meters:
              type: number
              example: 5000
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        face_person_group_id:
          type: string
          example: "session-uuid-person-group"
        total_votes:
          type: number
          example: 0
        results_published:
          type: boolean
          example: false
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        position:
          type: string
          example: "President"
        max_selections:
          type: number
          example: 1
        candidates:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
    Candidate:
      type: object
      properties:
        _id:
          type: string
        matric_no:
          type: string
          example: "BU22CSC1005"
        student_id:
          type: string
        manifesto:
          type: string
          example: "Building a united and progressive student community"
        votes:
          type: number
          example: 0
    Vote:
      type: object
      properties:
        _id:
          type: string
        student_id:
          type: string
        session_id:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              position:
                type: string
              candidate_id:
                type: string
        verification:
          type: object
          properties:
            face_match:
              type: boolean
            confidence:
              type: number
            location_verified:
              type: boolean
            distance_from_center:
              type: number
        voted_at:
          type: string
          format: date-time
paths:
...
  /sessions:
    get:
      tags:
        - Student - Sessions
      summary: List eligible sessions
      description: |
        Get all voting sessions that the authenticated student is eligible for
        based on their college, department, and level.
      operationId: listEligibleSessions
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
          description: Filter by session status
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/VotingSession'
        '401':
          description: Unauthorized
  /sessions/{id}:
    get:
      tags:
        - Student - Sessions
      summary: Get session details
      description: Get detailed information about a specific voting session
      operationId: getSession
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  session:
                    $ref: '#/components/schemas/VotingSession'
                  has_voted:
                    type: boolean
                    example: false
        '401':
          description: Unauthorized
        '404':
          description: Session not found

  # ==================== STUDENT - VOTING ====================
  /vote:
    post:
      tags:
        - Student - Voting
      summary: Submit vote
      description: |
        Submit a vote with facial verification and geofencing validation.
        
        ## Verification Steps:
        1. ✅ Session is active
        2. ✅ Student is eligible (college/department/level match)
        3. ✅ Student hasn't voted yet
        4. ✅ Student is within campus radius (5000m from center)
        5. ✅ Face matches enrolled photo (Azure Face API - 70% confidence minimum)
        
        **Success**: Sends vote confirmation email
      operationId: submitVote
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - session_id
                - choices
                - image_url
                - lat
                - lng
              properties:
                session_id:
                  type: string
                  example: "507f1f77bcf86cd799439011"
                choices:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - position
                      - candidate_id
                    properties:
                      position:
                        type: string
                        example: "President"
                      candidate_id:
                        type: string
                        example: "507f1f77bcf86cd799439012"
                  example:
                    - position: "President"
                      candidate_id: "507f1f77bcf86cd799439012"
                    - position: "Vice President"
                      candidate_id: "507f1f77bcf86cd799439013"
                image_url:
                  type: string
                  format: uri
                  example: "https://example.com/student-face.jpg"
                  description: "URL to student's face photo for verification"
                lat:
                  type: number
                  minimum: -90
                  maximum: 90
                  example: 7.8525
                  description: "Student's current latitude"
                lng:
                  type: number
                  minimum: -180
                  maximum: 180
                  example: 4.2811
                  description: "Student's current longitude"
      responses:
        '201':
          description: Vote submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Vote submitted successfully"
                  vote:
                    $ref: '#/components/schemas/Vote'
        '400':
          description: Validation error or voting not allowed
        '401':
          description: Unauthorized
        '403':
          description: Face verification failed or geofencing failed
        '429':
          description: Rate limit exceeded
  /vote/history:
    get:
      tags:
        - Student - Voting
      summary: Get voting history
      description: Get authenticated student's complete voting history across all sessions
      operationId: getVotingHistory
      security:
        - BearerAuth: []
      responses:
        '200':
          description: History retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  votes:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id:
                          type: string
                        session_title:
                          type: string
                        voted_at:
                          type: string
                          format: date-time
                        verification_status:
                          type: object
        '401':
          description: Unauthorized

  # ==================== RESULTS ====================
...